% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{contextToBinData}
\alias{contextToBinData}
\title{Converting data to binary table}
\usage{
contextToBinData(x, nQuantiles = 3)
}
\arguments{
\item{x}{is the data.frame that contains the data}

\item{nQuantiles}{the number of quantiles for binarization}
}
\value{
the new data.frame with binary values
}
\description{
The function binarize data in requested number of quantiles
taken into account that some quantiles can have a lot of observations
}
\examples{

}
