% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{toFCAPSJson}
\alias{toFCAPSJson}
\title{A function that writes a logical table as a formal context}
\usage{
toFCAPSJson(db, file)
}
\arguments{
\item{db}{is the dataset to code as a formal context}

\item{file}{-- is the file name to write the data to}
}
\description{
A function that writes a logical table as a formal context
}
\examples{

}
